test(http2): move GH-30342 coverage to test/regression/issue/30342.test.ts
The tests previously lived at the end of test/js/node/http2/node-http2.test.js
but that file has pre-existing flakes in its node-environment shards
that fail independently of this change. A dedicated regression file
makes the test set self-contained and the gate's signal legible.
